Crate re_viewport

Rerun Viewport Panel

This crate provides the central panel that contains all space views.

Modules

auto_layout 🔒

Code for automatic layout of space views.

blueprint_components

Potentially user-facing components to be used in blueprints.

external

space_info 🔒

space_view 🔒

space_view_entity_picker 🔒

space_view_heuristics 🔒

space_view_highlights 🔒

viewport 🔒

The viewport panel.

viewport_blueprint 🔒

viewport_blueprint_ui 🔒

Structs

SpaceInfoCollection

Information about all spaces.

SpaceViewBlueprint

A view of a space.

Viewport

Defines the layout of the Viewport

ViewportBlueprint

Describes the layout and contents of the Viewport Panel.

ViewportState

State for the Viewport that persists across frames but otherwise is not saved.

Functions

query_pinhole 🔒

Utility for querying a pinhole archetype instance.