Module re_viewport::space_view_heuristics

Functions

all_possible_space_views

List out all space views we allow the user to create.

candidate_space_view_paths 🔒

contains_any_image 🔒

default_created_space_views

List out all space views we generate by default for the available data.

identify_entities_per_system_per_class

is_entity_processed_by_class

is_entity_processed_by_part_collection 🔒

Returns true if an entity is processed by any of the given re_viewer_context::ViewPartSystems.

is_interesting_space_view_at_root 🔒

is_interesting_space_view_not_at_root 🔒

is_spatial_class 🔒

Returns true if a class is one of our spatial classes.

reachable_entities_from_root

spawn_one_space_view_per_entity 🔒

Type Definitions

EntitiesPerSystem

EntitiesPerSystemPerClass