Function re_viewport::space_view_highlights::highlights_for_space_view

pub fn highlights_for_space_view(
    ctx: &ViewerContext<'_>,
    space_view_id: SpaceViewId
) -> SpaceViewHighlights

Computes which things in a space view should received highlighting.

This method makes decisions which entities & instances should which kind of highlighting based on the entities in a space view and the current selection/hover state.