Struct re_viewport::viewport::TabViewer

struct TabViewer<'a, 'b> {
    viewport_state: &'a mut ViewportState,
    ctx: &'a mut ViewerContext<'b>,
    space_views: &'a mut BTreeMap<SpaceViewId, SpaceViewBlueprint>,
    maximized: &'a mut Option<SpaceViewId>,
    edited: bool,
}

egui_tiles has tiles which are either containers or panes.

In our case, each pane is a space view, while containers are just groups of things.

Fields
viewport_state: &'a mut ViewportStatectx: &'a mut ViewerContext<'b>space_views: &'a mut BTreeMap<SpaceViewId, SpaceViewBlueprint>maximized: &'a mut Option<SpaceViewId>edited: bool

The user edited the tree.