Struct re_viewport::viewport::TabViewer

struct TabViewer<'a, 'b> {
    viewport_state: &'a mut ViewportState,
    ctx: &'a ViewerContext<'b>,
    viewport_blueprint: &'a ViewportBlueprint,
    maximized: &'a mut Option<SpaceViewId>,
    root_container_id: Option<ContainerId>,
    tree_action_sender: Sender<TreeAction>,
    executed_systems_per_space_view: HashMap<SpaceViewId, (ViewQuery<'a>, SystemExecutionOutput)>,
    contents_per_tile_id: HashMap<TileId, Contents>,
    edited: bool,
}

egui_tiles has tiles which are either containers or panes.

In our case, each pane is a space view, while containers are just groups of things.

Fields
viewport_state: &'a mut ViewportStatectx: &'a ViewerContext<'b>viewport_blueprint: &'a ViewportBlueprintmaximized: &'a mut Option<SpaceViewId>root_container_id: Option<ContainerId>tree_action_sender: Sender<TreeAction>executed_systems_per_space_view: HashMap<SpaceViewId, (ViewQuery<'a>, SystemExecutionOutput)>

List of query & system execution results for each space view.

contents_per_tile_id: HashMap<TileId, Contents>

List of contents for each tile id

edited: bool

The user edited the tree.